orchestrator/requirements.txt
fastapi
uvicorn
opencensus-ext-azure
